netz: it works!  correctly filtering out particular field.
basically we keep track of where we are in the output; if the
desired field is also there, we push it out.

also started using my debug code here too.. if the code for the
NETZ guys is going to live in this repo, it might as well use it.
but we should probably move the code to be separate at some
point...