netz: parse field options.
parse field options in psiphi.cfg.  the two options now are "3D" and
"slice=Z"; "3D" is the old behavior of just writing out the full
3D field.  "slice=Z" is not yet implemented, but is planned to
pull out a specific slice.

also switch up 'byteintersect' to not take a field, but rather the
field's upper && lower boundaries.  these are the only parts of
the field it needed; this makes it more general.