netz: write out slice information if requested.
instead of hacking it, like we do in the 3D case, this is The Real
Deal.  We don't copy it to a file first and then pull out what we
need; we actually identify what we need and push it out directly
from the incoming stream.

the implementation feels waaaay too complicated, and thus i'm not
a big fan of it.  I also don't think this will work if there are
multiple bricks in any dimension sans X.

i also converted slicefield[] to an array of descriptors, because
we use lio_listio to submit all the writes in one batch.