optimizations to the MC implementation.
do the loops in the correct order.  switch to using two pointers
for the top and bottom, instead of calculating where is on its
own.  hopefully in the future this simplifies buffer management.
other small optimizations.  delete dead code.  switch to binary
i/o instead of printf nonsense... which means we can't load it in
meshlab directly, unfortunately.

these optimizations give it a 10x speedup, which makes it 'only'
10x behind what transport without isosurfaces results in.  sigh.