grid_size -> metadata
figure out the datatype from hdf5... if that type is 64-bit
floating point, for now.  Turns out reading the type from hdf5 is
a nightmare.

but set up all the infrastructure so we can do this generically.
'grid_size' got renamed to 'metadata', since it now communicates
type information as well.